requirements.txt
Celery==4.0.2
python-dateutil==2.6.0
Django==1.10.6
dj-database-url==0.4.2
djangorestframework==3.6.2
django-bootstrap3==8.2.1
django-celery-beat==1.0.1
django-celery-results==1.0.1
django-debug-toolbar==1.7
django-extensions==1.7.8
django-oauth-toolkit==0.12.0
factory-boy==2.8.1
gunicorn==19.7.1
raven==6.0.0
requests==2.13.0
psycopg2==2.7.1
scrapy==1.3.3
whitenoise==3.3.0
